WebDetailsCache class

WebDetailsCache class

WebDetailsCache class

Details

A class for caching web details including keys, metadata, and entry details.

Public fields

cache_name

Unique name for the cache.

cache_dir

Directory for cache files.

log_dir

Directory for log files.

all_entry_keys

List of cached entry keys.

all_entry_details

List of cached entry details.

metadata

Metadata associated with the cache.

list_all_entries_url

URL for listing all entries.

key_field_extraction_func

Function to extract the key from entries.

single_entry_details_url

URL template for fetching single entry details.

all_entry_keys_file

File path for storing entry keys.

metadata_file

File path for storing metadata.

all_entry_details_file

File path for storing entry details.

flush_threshold

Number of entry detils fetched before flushing to disk. (improves perf since writing to disk can take a long time)

list_all_entries_page_size

Page size for listing entries.

Methods

Public methods

WebDetailsCache$new()

WebDetailsCache$load_cache()

WebDetailsCache$build_cache()

WebDetailsCache$are_all_entry_keys_cached()

WebDetailsCache$are_all_entry_details_cached()

WebDetailsCache$fetch_all_entry_keys()

WebDetailsCache$fetch_all_entry_details()

WebDetailsCache$flush_all_entry_keys()

WebDetailsCache$flush_all_entry_details()

WebDetailsCache$clone()

Method new()

initialize a new WebDetailsCache

Usage

WebDetailsCache$new(
  cache_name,
  cache_dir,
  log_dir,
  list_all_entries_url,
  key_field_extraction_func,
  single_entry_details_url,
  flush_threshold = 100,
  list_all_entries_page_size = 1000
)

Arguments

cache_name

Unique name for the cache.

cache_dir

Directory to store cache files.

log_dir

Directory for log files.

list_all_entries_url

URL for listing all entries.

key_field_extraction_func

Function to extract the key from entries.

single_entry_details_url

URL template for fetching details for a single entry.

flush_threshold

flush_threshold Number of entry detils fetched before flushing to disk (default 100).

list_all_entries_page_size

Number of entries per page (default 1000).

Method load_cache()

Load the cache data from disk

Loads cached keys, metadata, and entry details from their respective files if they exist.

Usage

WebDetailsCache$load_cache()

Method build_cache()

Build the cache.

Loads existing cache data and fetches missing entry keys and details.

Usage

WebDetailsCache$build_cache()

Method are_all_entry_keys_cached()

Check if all entry keys are cached.

Usage

WebDetailsCache$are_all_entry_keys_cached()

Returns

TRUE if the number of cached keys equals the total number of entries; FALSE otherwise.

Method are_all_entry_details_cached()

Check if all entry details are cached.

Usage

WebDetailsCache$are_all_entry_details_cached()

Returns

TRUE if every key has corresponding details cached; FALSE otherwise.

Method fetch_all_entry_keys()

Fetch all entry keys from the web.

Retrieves entry keys in a paginated manner and updates the cache.

Usage

WebDetailsCache$fetch_all_entry_keys()

Method fetch_all_entry_details()

Fetch all entry details from the web.

Iterates over all entry keys and fetches detailed information for each.

Usage

WebDetailsCache$fetch_all_entry_details()

Method flush_all_entry_keys()

Flush cached entry keys to disk.

Atomically saves the cached keys and metadata to disk.

Usage

WebDetailsCache$flush_all_entry_keys()

Method flush_all_entry_details()

Flush cached entry details to disk.

Atomically saves the cached entry details to disk.

Usage

WebDetailsCache$flush_all_entry_details()

Method clone()

The objects of this class are cloneable with this method.

Usage

WebDetailsCache$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.