Update DESCRIPTION file with package dependencies

desc_pkgs will search code in a package for dependencies, using both renv::dependencies to search R and RMD files, and by detecting dependencies declared with roxygen2 decorators. It will write these package dependencies to the DESCRIPTION file.

desc_pkgs(
  dir = ".",
  file = "DESCRIPTION",
  suggests = c("^inst/"),
  ignore = c("^data-raw/"),
  must_keep = character(0),
  quiet = TRUE
)

Arguments

dir

Root package directory

file

DESCRIPTION file name

suggests

Path(s) to search for dependencies to put in SUGGESTS (supports regular expressions)

ignore

Path(s) to ignore in getting dependencies (supports regular expressions)

must_keep

List of packages that must be retained even if not a current dependency

quiet

Whether to print messages when renv::dependencies is called