Create multi config file

Prepare the cellranger multi file for some samples.

Usage

llrs_cellranger_multi(config, gex, library, samples, vdj, feature, out)

Arguments

config

A list with the general config of the samples.

gex

A data.frame with the information for the gene expression (GEX).

library

A data.frame with the information about the libraries used.

samples

A data.frame with the information about the samples used.

vdj

A list with the configuration for the vdj assays.

feature

A list with the configuration for the featured assays.

out

Path to a directory where to save the file.

Value

The path to a file saved with the configuration as specified.

References

https://www.10xgenomics.com/support/software/cell-ranger/latest/advanced/cr-multi-config-csv-opts

See also

Examples

#llrs_cellranger_multi(out = "/S2_cellranger.csv")