Swap values in a vector

swap_vector() swaps values in a vector with a specified value.

Usage

swap_vector(x, swap = na_as(x), except = NULL, preserve_nas = FALSE)

Arguments

x

An atomic vector.

swap

(Optional) A single value to swap the values in x (Default: na_as(x)).

except

(Optional) A vector of values to exclude from swapping.

preserve_nas

(Optional) A logical flag to indicating if the function must preserve NA values in x (Default: FALSE).

Value

The same type of object in x with swapped values.

See also

Examples

swap_vector(1:10)
#>  [1] NA NA NA NA NA NA NA NA NA NA
#> [1] NA NA NA NA NA NA NA NA NA NA

swap_vector(1:10, swap = 0L, except = 5:7)
#>  [1] 0 0 0 0 5 6 7 0 0 0
#> [1] 0 0 0 0 5 6 7 0 0 0

swap_vector(c("a", "b", NA, "c"), swap = "d", preserve_nas = TRUE)
#> [1] "d" "d" NA  "d"
#> [1] "d" "d" NA  "d"