Update package versions in DESCRIPTION file

update_pkg_versions() updates the version of the packages in the DESCRIPTION file of R packages with their current version.

If the package is an R base package, the function will update the version to the previous minor version of the current R version. This is made to avoid errors with CI/CD.

update_pkg_versions(
  file = here::here("DESCRIPTION"),
  old_r_version = bump_back_r_version()
)

Arguments

file

(optional) A string indicating the path to the DESCRIPTION file (default: here::here("DESCRIPTION")).

old_r_version

(optional) A string indicating the previous minor version of the current R version (default: bump_back_r_version()).

Value

An invisible NULL. This function is used for its side effect.

See also