Set of functions to setup default R and git configuration files

These functions are used to setup default .Rprofile/.Renviron/.gitconfig/.gitignore_global/.rstudio-prefs.json configurations. This settings are based on my preference and some files, e.g. .gitconfig contain info that will be specific to me personally. By default, all functions will overwrite the existing files, though this can be modified through the append argument.

setup_r_git(append = FALSE)

setup_Rprofile(
  path = "~/.Rprofile",
  template = "template_Rprofile",
  append = FALSE
)

setup_Renviron(
  path = "~/.Renviron",
  template = "template_Renviron",
  append = FALSE
)

setup_gitconfig(
  path = "~/.gitconfig",
  template = "template_gitconfig",
  append = FALSE
)

setup_gitignore_global(
  path = "~/.gitignore_global",
  template = "template_gitignore_global",
  append = FALSE
)

setup_rstudio_prefs(
  path = "~/.config/rstudio/rstudio-prefs.json",
  template = "template_rstudio-prefs",
  append = FALSE
)

Arguments

append

logical(1) whether to append to or overwrite the path with the template.

path

character(1) path to the configuration file.

template

character(1) name of the template configuration files stored in inst/extdata

Functions

setup_Rprofile: Set up a default ~/.Rprofile

setup_Renviron: Set up a default ~/.Renviron

setup_gitconfig: Set up a default ~/.gitconfig

setup_gitignore_global: Set up a default ~/.gitignore_global

setup_rstudio_prefs: Set up a default ~/.config/rstudio/rstudio-prefs.json

See also

Examples

if (FALSE) {
setup_Rprofile()
setup_Renviron()
setup_gitconfig()
setup_gitignore_global()
setup_rstudio_prefs()
}